refactor: policy-studio - move into dedicated directory
