+++ projects/ui-policy-studio-angular/src/lib/components/flow-details-info-bar/gio-ps-flow-details-info-bar.component.ts
@@ -16,7 +16,7 @@
 import { Component, Input } from '@angular/core';
 import { isEmpty } from 'lodash';
 
-import { FlowVM } from '../../gio-policy-studio.model';
+import { FlowVM } from '../../policy-studio/gio-policy-studio.model';
 import { ChannelSelector, ConditionSelector, ConnectorInfo, HttpSelector, Operation } from '../../models';
 
 @Component({
