+++ projects/ui-policy-studio-angular/src/lib/components/flow-execution-form-dialog/gio-ps-flow-execution-form-dialog.stories.ts
@@ -20,7 +20,7 @@ import { tap } from 'rxjs/operators';
 import { action } from '@storybook/addon-actions';
 import { GIO_DIALOG_WIDTH } from '@gravitee/ui-particles-angular';
 
-import { GioPolicyStudioModule } from '../../gio-policy-studio.module';
+import { GioPolicyStudioModule } from '../../policy-studio/gio-policy-studio.module';
 import { FlowExecution } from '../../models';
 import { fakeBestMatchFlowExecution } from '../../models/flow/FlowExecution.fixture';
