+++ projects/ui-policy-studio-angular/src/lib/components/flow-form-dialog/flow-message-form-dialog/gio-ps-flow-message-form-dialog.stories.ts
@@ -20,8 +20,8 @@ import { tap } from 'rxjs/operators';
 import { action } from '@storybook/addon-actions';
 import { GIO_DIALOG_WIDTH } from '@gravitee/ui-particles-angular';
 
-import { GioPolicyStudioModule } from '../../../gio-policy-studio.module';
-import { FlowVM } from '../../../gio-policy-studio.model';
+import { GioPolicyStudioModule } from '../../../policy-studio/gio-policy-studio.module';
+import { FlowVM } from '../../../policy-studio/gio-policy-studio.model';
 import { fakeChannelFlow, fakeHTTPGetMessageEntrypoint, fakeSSEMessageEntrypoint } from '../../../models/index-testing';
 import { GioPolicyStudioFlowFormDialogResult } from '../gio-ps-flow-form-dialog-result.model';
