+++ projects/ui-policy-studio-angular/src/lib/components/flow-form-dialog/flow-proxy-form-dialog/gio-ps-flow-proxy-form-dialog.component.spec.ts
@@ -23,9 +23,9 @@ import { MatButtonHarness } from '@angular/material/button/testing';
 import { InteractivityChecker } from '@angular/cdk/a11y';
 import { MatIconTestingModule } from '@angular/material/icon/testing';
 
-import { GioPolicyStudioModule } from '../../../gio-policy-studio.module';
+import { GioPolicyStudioModule } from '../../../policy-studio/gio-policy-studio.module';
 import { fakeChannelFlow, fakeHttpFlow } from '../../../models/index-testing';
-import { FlowVM } from '../../../gio-policy-studio.model';
+import { FlowVM } from '../../../policy-studio/gio-policy-studio.model';
 import { GioPolicyStudioFlowFormDialogResult } from '../gio-ps-flow-form-dialog-result.model';
 import { Flow, HttpMethod, Operator, toHttpMethod } from '../../../models';
