+++ projects/ui-policy-studio-angular/src/lib/components/step-edit-dialog/gio-ps-step-edit-dialog.component.spec.ts
@@ -27,9 +27,9 @@ import { MatInputHarness } from '@angular/material/input/testing';
 import { HttpClientTestingModule } from '@angular/common/http/testing';
 
 import { fakeTestPolicyStep, fakeTestPolicy } from '../../models/index-testing';
-import { GioPolicyStudioModule } from '../../gio-policy-studio.module';
+import { GioPolicyStudioModule } from '../../policy-studio/gio-policy-studio.module';
 import { Policy, Step } from '../../models';
-import { GioPolicyStudioService } from '../../gio-policy-studio.service';
+import { GioPolicyStudioService } from '../../policy-studio/gio-policy-studio.service';
 import { fakePolicyDocumentation, fakePolicySchema } from '../../models/policy/PolicySchema.fixture';
 
 import {
