+++ projects/ui-policy-studio-angular/src/lib/policy-studio/gio-policy-studio.component.html
