+++ projects/ui-policy-studio-angular/src/lib/policy-studio/gio-policy-studio.component.scss
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-@use '@angular/material' as mat;
+@use '../../../../../node_modules/@angular/material/index' as mat;
 @use '@gravitee/ui-particles-angular' as gio;
 
 :host {
