+++ projects/ui-policy-studio-angular/src/public-testing-api.ts
@@ -15,4 +15,4 @@
  */
 export * from './lib/models/index-testing';
 
-export * from './lib/gio-policy-studio.harness';
+export * from './lib/policy-studio/gio-policy-studio.harness';
